requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
geopandas>=0.14.0
folium>=0.15.0
streamlit-folium>=0.15.0
xarray>=2023.1.0
netCDF4>=1.6.0
h5netcdf>=1.0.0
cdsapi>=0.6.1
requests>=2.31.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
statsmodels>=0.14.0
rasterio>=1.3.0
shapely>=2.0.0
plotly>=5.17.0
openpyxl>=3.1.0
python-dotenv>=1.0.0